refactor: added operations folder to work with createAsyncThunk, created a reusable component
